split LoginPanel into LoginForm and AccountPanel
