+++ web/resources/lib/account-panel.ts
@@ -1,0 +1,4 @@
+import { Modal } from "@wikijump/components"
+import AccountPanel from "./components/AccountPanel.svelte"
+
+export const LoginModal = new Modal(AccountPanel, true)
