--- web/resources/lib/login-modal.ts
@@ -1,4 +1,0 @@
-import { Modal } from "@wikijump/components"
-import LoginPanel from "./components/LoginPanel.svelte"
-
-export const LoginModal = new Modal(LoginPanel, true)
