+++ web/resources/scripts/index.ts
@@ -1,5 +1,5 @@
 import "@wikijump/ftml-components"
+import "../lib/account-panel"
 import "../lib/elements/nav-dropdown"
-import "../lib/login-modal"
 
 export {}
